Cast for the second argument to pow() to satisfy gcc33.
git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@7745 26435e26-4b25-483a-942d-477236992e84

+++ trunk/xorp/bgp/damping.cc
@@ -96,7 +96,7 @@ Damping::init()
     for (size_t i = 0; i < array_size; i++) {
 	_decay[i] = static_cast<uint32_t>(decay_i * FIXED);
 // 	printf("%d %d %f\n",i, _decay[i], decay_i);
-	decay_i = pow(decay_1, i + 2);
+	decay_i = pow(decay_1, static_cast<int>(i + 2));
     }
 
     // Start the timer to incement the tick
